% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.petrinet.R
\name{as.petrinet}
\alias{as.petrinet}
\title{Converts the object to a Petrinet}
\usage{
as.petrinet(obj)
}
\arguments{
\item{obj}{The event log to be used. An object of class}
}
\description{
Converts the object to a Petrinet
}
\examples{
library(petrinetR)
cn <- causal_net(L_heur_1, threshold = .8)
pn <- as.petrinet(cn)
print(pn)
render_PN(pn)

}
