% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_frequency.R
\name{causal_frequency}
\alias{causal_frequency}
\title{Frequency map profile}
\usage{
causal_frequency(
  value = c("absolute", "relative"),
  color_scale = "PuBu",
  color_edges = "dodgerblue4"
)
}
\arguments{
\item{value}{The type of frequency value to be used:
absolute, relative (percentage of activity instances).}

\item{color_scale}{Name of color scale to be used for nodes. Defaults to PuBu. See \code{Rcolorbrewer::brewer.pal.info()} for all options.}

\item{color_edges}{The color used for edges. Defaults to dodgerblue4.}
}
\description{
Function to create a frequency profile for a process map.
}
\examples{
causal_net(L_heur_1,
           type = causal_frequency("relative"))

}
