% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics_map.R
\name{dependency_graph}
\alias{dependency_graph}
\title{Create a dependency graph}
\usage{
dependency_graph(dependencies, rankdir = "LR", layout = "dot",
  render = T)
}
\arguments{
\item{dependencies}{A dependency matrix created by \code{\link{dependency_matrix}}}

\item{rankdir}{Rankdir to be used for DiagrammeR.}

\item{layout}{Layout to be used for DiagrammeR.}

\item{render}{Whether to directly render the DiagrammeR graph or simply return it.}
}
\value{
A DiagrammeR graph of the (filtered) dependency matrix.
}
\description{
Creates a dependency graph visualizing a dependency matrix and
}
\examples{
library(eventdataR)
library(processmapR)
data(patients)
dependency_graph(dependency_matrix(patients))

}
