% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_type_fhm.R
\name{dependency_type_fhm}
\alias{dependency_type_fhm}
\title{Flexible heuristics miner dependency type}
\usage{
dependency_type_fhm(threshold = 0.9, threshold_l1 = threshold,
  threshold_l2 = threshold, all_connected = FALSE, precedence = NULL,
  precedence_lenght_two_loops = NULL)
}
\arguments{
\item{threshold}{A dependency threshold, usually in the interval \code{[0,1]}, filtering out dependencies below the threshold.}

\item{threshold_l1}{A dependency threshold, usually in the interval \code{[0,1]}, filtering out self-loop dependencies below the threshold.}

\item{threshold_l2}{A dependency threshold, usually in the interval \code{[0,1]}, filtering out length-two loop dependencies below the threshold.}

\item{all_connected}{If \code{TRUE} the best antecedent and consequent (as determined by the dependency measure) are going to be added regardless of the \code{threshold} value.}

\item{precedence}{A precedence matrix overriding the default matrix generated from the event log. For example, created from a bupaR event log using (\code{\link{precedence_matrix_absolute}}).}

\item{precedence_lenght_two_loops}{A precedence matrix of length two loop patterns (aba) overriding the default matrix generated from the event log. For example, created from a bupaR event log using (\code{\link{precedence_matrix_length_two_loops}}).}
}
\value{
A dependency type.
}
\description{
Flexible heuristics miner dependency type
}
\examples{
dependency_matrix(L_heur_1,
                  dependency_type = dependency_type_fhm(all_connected = TRUE))

}
