% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{L_heur_1}
\alias{L_heur_1}
\title{Heuristics miner example log #1}
\format{
Eventlog containing 40  cases
}
\source{
\url{http://www.processmining.org/event_logs_and_models_used_in_book}
}
\usage{
L_heur_1
}
\description{
Artificial eventlog for illustrating Heuristics Miner published as supplementary material to the book
\verb{Process Mining: Discovery, Conformance and Enhancement of Business Processes}.
}
\references{
Process Mining: Discovery, Conformance and Enhancement of Business Processes by W.M.P. van der Aalst, Springer Verlag, 2011 (ISBN 978-3-642-19344-6).
}
\keyword{datasets}
