% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parallel_matrix_lifecycle.R
\name{parallel_matrix_lifecycle}
\alias{parallel_matrix_lifecycle}
\title{Parallel Matrix with Lifecycle}
\usage{
parallel_matrix_lifecycle(eventlog)
}
\arguments{
\item{eventlog}{The event log object to be used.}
}
\description{
Parallel Matrix with Lifecycle
}
\examples{
parallel_matrix_lifecycle(L_heur_1)

}
