% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.dependency_matrix.R
\name{plot.dependency_matrix}
\alias{plot.dependency_matrix}
\title{Dependency matrix plot}
\usage{
\method{plot}{dependency_matrix}(x, ...)
}
\arguments{
\item{x}{Dependency matrix}

\item{...}{Additional parameters}
}
\value{
A ggplot object, which can be customized further, if deemed necessary.
}
\description{
Visualize a dependency matrix. A generic plot function for dependency matrices.
}
