% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precedence_matrix_absolute.R
\name{precedence_matrix_absolute}
\alias{precedence_matrix_absolute}
\title{Precendence Matrix}
\usage{
precedence_matrix_absolute(eventlog)
}
\arguments{
\item{eventlog}{The event log object to be used}
}
\description{
Construct a precendence matrix, showing how activities are followed by each other. This function computes the precedence matrix directly in C++ for efficiency.
Only the type \code{absolute} of (\code{\link[processmapR]{precedence_matrix}}) is supported.
}
\examples{
\dontrun{
library(eventdataR)
data(traffic_fines)
precedence_matrix_absolute(traffic_fines)
}

}
