% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precedence_matrix_length_two_loop.R
\name{precedence_matrix_length_two_loops}
\alias{precedence_matrix_length_two_loops}
\title{Length Two Loop Precedence Matrix}
\usage{
precedence_matrix_length_two_loops(eventlog)
}
\arguments{
\item{eventlog}{The event log object to be used.}
}
\description{
Construct a precedence matrix counting how often pattern \code{aba} occurs.
}
\examples{
\dontrun{
library(eventdataR)
data(traffic_fines)
m <- precedence_matrix_length_two_loop(traffic_fines)
print(m)
as.matrix(m)
}

}
