% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.causal_net.R
\name{print.causal_net}
\alias{print.causal_net}
\title{Generic print function for a Causal net}
\usage{
\method{print}{causal_net}(x, ...)
}
\arguments{
\item{x}{Causal net object}

\item{...}{Additional Arguments}
}
\description{
Generic print function for a Causal net
}
