% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.dependency_matrix.R
\name{print.dependency_matrix}
\alias{print.dependency_matrix}
\title{Generic print function for a dependency matrix}
\usage{
\method{print}{dependency_matrix}(x, ...)
}
\arguments{
\item{x}{dependency matrix object}

\item{...}{Additional Arguments}
}
\description{
Generic print function for a dependency matrix
}
