% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_dependency_matrix.R
\name{render_dependency_matrix}
\alias{render_dependency_matrix}
\title{Renders a dependency matrix as dependency graph}
\usage{
render_dependency_matrix(
  dependencies,
  rankdir = "LR",
  layout = "dot",
  render = T
)
}
\arguments{
\item{dependencies}{A dependency matrix created by \code{\link{dependency_matrix}}}

\item{rankdir}{Rankdir to be used for DiagrammeR.}

\item{layout}{Layout to be used for DiagrammeR.}

\item{render}{Whether to directly render the DiagrammeR graph or simply return it.}
}
\value{
A DiagrammeR graph of the (filtered) dependency matrix.
}
\description{
Creates a dependency graph visualizing the contents of a dependency matrix.
}
\examples{
render_dependency_matrix(dependency_matrix(L_heur_1))

}
