% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_map.R
\name{causal_map}
\alias{causal_map}
\title{Create a causal map}
\usage{
causal_map(processmap, dependencies, rankdir = "LR", layout = "dot",
  render = T)
}
\arguments{
\item{processmap}{A proces map created by processmapR.}

\item{dependencies}{A dependency matrix created for the event log, for example, by \code{\link{dependency_matrix}}.}

\item{rankdir}{Rankdir to be used for DiagrammeR.}

\item{layout}{Layout to be used for DiagrammeR.}

\item{render}{Whether to directly render the DiagrammeR graph or simply return it.}
}
\value{
A DiagrammeR graph of the causal map.
}
\description{
Creates a causal map which is similar to a processmapR process map. However, the causal map
deals with parallelism by trying to identifying causal dependencies between activities using
several heuristics.
}
\details{
Warning: Projected frequencies are WORK IN PROGRESS and currently not reliable!
}
\examples{
library(eventdataR)
library(processmapR)
data(patients)
causal_map(patients, dependency_matrix(patients))

}
