% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heuristics_dependencies.R
\name{dependency_matrix}
\alias{dependency_matrix}
\title{Create a dependency matrix}
\usage{
dependency_matrix(eventlog = NULL,
  precedence = precedence_matrix(eventlog, type = "absolute"),
  threshold = 0, type = c("FHM"))
}
\arguments{
\item{eventlog}{A bupaR event log, may be NULL when a precedence matrix is provided.}

\item{precedence}{A precedence matrix, for example, created from a bupaR event log using (\code{\link{precedence_matrix}}).}

\item{threshold}{A dependency threshold in the interval \code{[0,1]} filtering out dependencies below the threshold.}

\item{type}{Which approach to use for calculation of the dependency matrix. Currently only \code{FHM} available.}
}
\value{
A data frame with the dependency values between activities.
}
\description{
Creates a dependency matrix from a precedence matrix (\code{\link{precedence_matrix}}) based on different approaches.
}
\examples{
library(eventdataR)
library(processmapR)
data(patients)
dependency_matrix(patients)

}
\seealso{
\code{\link{precedence_matrix}}, \code{\link{dependency_graph}}
}
