% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependency_matrix.R
\name{dependency_matrix}
\alias{dependency_matrix}
\title{Create a dependency matrix}
\usage{
dependency_matrix(eventlog = NULL,
  precedence = precedence_matrix_absolute(eventlog), threshold = 0.9,
  type = c("FHM"))
}
\arguments{
\item{eventlog}{A bupaR event log, may be NULL when a precedence matrix is provided.}

\item{precedence}{A precedence matrix as data frame, for example, created from a bupaR event log using (\code{\link{precedence_matrix_absolute}}).}

\item{threshold}{A dependency threshold in the interval \code{[0,1]} filtering out dependencies below the threshold.}

\item{type}{Which approach to use for calculation of the dependency matrix. Currently only \code{FHM} available.}
}
\value{
A data frame with class \code{dependency_matrix} containing the computed dependency values between activities.
}
\description{
Creates a dependency matrix from a precedence matrix (\code{\link{precedence_matrix}}) based on different approaches.
}
\examples{
d <- dependency_matrix(L_heur_1)
print(d)
as.matrix(d)

}
\seealso{
\code{\link{precedence_matrix}}
}
