% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/precedence_matrix_lifecycle.R
\name{precedence_matrix_lifecycle}
\alias{precedence_matrix_lifecycle}
\title{Precedence Matrix with Lifecycle}
\usage{
precedence_matrix_lifecycle(eventlog)
}
\arguments{
\item{eventlog}{The event log object to be used.}
}
\description{
Precedence Matrix with Lifecycle
}
\examples{
precedence_matrix_lifecycle(L_heur_1)

}
